Add bash profile and make install script more idempotent

+++ install.sh
@@ -1,9 +1,9 @@
 #!/bin/bash
 
-for dotfile in .bashrc .gitconfig .gitignore_global .vimrc; do
-  ln -s "$(pwd)/$dotfile" ~
+for dotfile in .bashrc .bash_profile .gitconfig .gitignore_global .vimrc; do
+  [ -e ~/$dotfile ] || ln -s "$(pwd)/$dotfile" ~
 done
 
 mkdir -p ~/.vim
-git clone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
-ln -s "$(pwd)/.vim/syntax" ~/.vim/syntax
+[ -e ~/.vim/bundle/Vundle.vim ] ||  git clone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
+[ -e ~/.vim/syntax ] || ln -s "$(pwd)/.vim/syntax" ~/.vim/syntax
